+++ docs/release_notes.rst
@@ -4,4 +4,5 @@ Release Notes
 v0.1
 ----
 
-* This is the initial release of ecs-transform.
+* This is the initial release of container-transform.
+* Includes a Fig to ECS transformer